agent/requirements.txt
openai>=1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot>=20.0
python-dotenv
schedule
regex
cryptography
openai>=1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
schedule
regex
cryptography
supabase>=2.0.0